Rename ground_bridge to bridged_ground
