+++ symbols/bridged_ground_left.ts
@@ -1,0 +1,4 @@
+import { flipSymbolOverYAxis } from "drawing/rotateSymbol"
+import ground_bridge_right from "./bridged_ground_right"
+
+export default flipSymbolOverYAxis(ground_bridge_right)
