+++ symbols/bridged_ground_up.ts
@@ -1,4 +1,4 @@
-import svgJson from "assets/generated/dual_n_channel_mosfet_transistor.json"
+import svgJson from "assets/generated/bridged_ground.json"
 import { modifySymbol } from "drawing/modify-symbol/modify-symbol"
 
 const { paths, texts, bounds, refblocks, circles } = svgJson
