--- symbols/dual_n_channel_mosfet_transistor_left.ts
@@ -1,4 +1,0 @@
-import { flipSymbolOverYAxis } from "drawing/rotateSymbol"
-import dual_n_channel_mosfet_transistor_right from "./dual_n_channel_mosfet_transistor_right"
-
-export default flipSymbolOverYAxis(dual_n_channel_mosfet_transistor_right)
